Derive seed for MurmurHash64A automatically
Given the hash of the empty string for the root directory entry we can
deterministically derive the seed used for MurmurHash64A.

Mathematically, this is possible due to how all three steps (xor, mul,
xor) of hashing an empty string with this algorithm are invertible.

As we don't know what algorithm could be used in the future the code
has a safeguard in that it validates the computed seed against the hash
of the next non-empty directory path rep.