feat: gather and create directories up-front
It turns out that creating the whole directory tree up to the root every
time a file is created is expensive even when the filesystem reports
no-ops for most parent directories due to the sheer volume of attempts.

This change gathers up the directories in a phase up-front and creates
all needed directories recursively from the target directory location.