feat: decompress bundles mostly into target buffer
When decompressing whole bundles we need to separately decompress each
contributing block into a buffer with additional safe space at the end.
When decompressing multiple blocks back to back, if the remaining space
in the target buffer is large to fit the safe space we can decompress
directly into the target buffer. This avoids a lot of allocations.