feat: use file I/O for GGPK data, avoid reindex
Memory-mapping a GGPK and touching most of its data seems to have a
noticeable memory load on Windows, consuming a lot of physical RAM pages
for data that is only read once. This is a larger problem with PoE2 as
that has over twice the amount of data as PoE1 and will just grow in the
future.

This change naively re-opens the pack file on each data file access and
reads out the bundle contents, leading to more I/O operations but less
persistent load on the file cache.

This also drive-by fixes a performance bug where it would index the
pack twice.